refactor: extract normalize_description into DescriptionNormalizable concern
Removes duplicated normalize_description method from CreativeCreateService
and CreativeUpdateService by extracting it into a shared
DescriptionNormalizable module.
